requirements.txt
elasticsearch<=7.0.0
neo4j>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0
numpy>=1.24.0
pandas>=2.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
pybloom_live